// frontend/utils/types.ts
export interface BaseEntity {
  id?: number | string;
  created_at?: string;
  updated_at?: string;
}
// ...
export interface AccountProps extends BaseEntity {
  name: string;
  account_number?: string;
  account_type?: string;
  account_status?: string;
  account_affects?: string;
  generates_income?: boolean;
}